Added handleChange functions for title and category and added them to jsx added comments for mapDispatchToProps
