+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"C:\\Users\\Rory\\Desktop\\Programming\\Microverse\\Main Curriculum\\React\\react-redux-bookstore\\src\\index.js":"1","C:\\Users\\Rory\\Desktop\\Programming\\Microverse\\Main Curriculum\\React\\react-redux-bookstore\\src\\utilities.js":"2","C:\\Users\\Rory\\Desktop\\Programming\\Microverse\\Main Curriculum\\React\\react-redux-bookstore\\src\\containers\\BooksList.js":"3","C:\\Users\\Rory\\Desktop\\Programming\\Microverse\\Main Curriculum\\React\\react-redux-bookstore\\src\\classes\\book.js":"4","C:\\Users\\Rory\\Desktop\\Programming\\Microverse\\Main Curriculum\\React\\react-redux-bookstore\\src\\types\\bookTypes.js":"5","C:\\Users\\Rory\\Desktop\\Programming\\Microverse\\Main Curriculum\\React\\react-redux-bookstore\\src\\components\\BooksForm.js":"6","C:\\Users\\Rory\\Desktop\\Programming\\Microverse\\Main Curriculum\\React\\react-redux-bookstore\\src\\components\\Book.js":"7"},{"size":347,"mtime":1611607563906,"results":"8","hashOfConfig":"9"},{"size":199,"mtime":1611355368657,"results":"10","hashOfConfig":"9"},{"size":987,"mtime":1611355800456,"results":"11","hashOfConfig":"9"},{"size":86,"mtime":1611355368657,"results":"12","hashOfConfig":"9"},{"size":84,"mtime":1611336840895,"results":"13","hashOfConfig":"9"},{"size":616,"mtime":1611607808920,"results":"14","hashOfConfig":"9"},{"size":449,"mtime":1611350321702,"results":"15","hashOfConfig":"9"},{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"14qb4vg",{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"26","messages":"27","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"28","messages":"29","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\Rory\\Desktop\\Programming\\Microverse\\Main Curriculum\\React\\react-redux-bookstore\\src\\index.js",[],"C:\\Users\\Rory\\Desktop\\Programming\\Microverse\\Main Curriculum\\React\\react-redux-bookstore\\src\\utilities.js",[],"C:\\Users\\Rory\\Desktop\\Programming\\Microverse\\Main Curriculum\\React\\react-redux-bookstore\\src\\containers\\BooksList.js",[],"C:\\Users\\Rory\\Desktop\\Programming\\Microverse\\Main Curriculum\\React\\react-redux-bookstore\\src\\classes\\book.js",[],"C:\\Users\\Rory\\Desktop\\Programming\\Microverse\\Main Curriculum\\React\\react-redux-bookstore\\src\\types\\bookTypes.js",[],"C:\\Users\\Rory\\Desktop\\Programming\\Microverse\\Main Curriculum\\React\\react-redux-bookstore\\src\\components\\BooksForm.js",[],"C:\\Users\\Rory\\Desktop\\Programming\\Microverse\\Main Curriculum\\React\\react-redux-bookstore\\src\\components\\Book.js",[]]
+[{"C:\\Users\\Rory\\Desktop\\Programming\\Microverse\\Main Curriculum\\React\\react-redux-bookstore\\src\\containers\\BooksForm.js":"1","C:\\Users\\Rory\\Desktop\\Programming\\Microverse\\Main Curriculum\\React\\react-redux-bookstore\\src\\index.js":"2"},{"size":1148,"mtime":1611615288089,"results":"3","hashOfConfig":"4"},{"size":325,"mtime":1611613919875,"results":"5","hashOfConfig":"6"},{"filePath":"7","messages":"8","errorCount":3,"warningCount":0,"fixableErrorCount":1,"fixableWarningCount":0,"source":null},"1c9hryo",{"filePath":"9","messages":"10","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"14qb4vg","C:\\Users\\Rory\\Desktop\\Programming\\Microverse\\Main Curriculum\\React\\react-redux-bookstore\\src\\containers\\BooksForm.js",["11","12","13"],"C:\\Users\\Rory\\Desktop\\Programming\\Microverse\\Main Curriculum\\React\\react-redux-bookstore\\src\\index.js",[],{"ruleId":"14","severity":2,"message":"15","line":14,"column":10,"nodeType":"16","messageId":"17","endLine":14,"endColumn":19},{"ruleId":"14","severity":2,"message":"18","line":15,"column":10,"nodeType":"16","messageId":"17","endLine":15,"endColumn":22},{"ruleId":"19","severity":2,"message":"20","line":43,"column":4,"nodeType":"21","messageId":"22","endLine":44,"endColumn":1,"fix":"23"},"no-unused-vars","'bookTitle' is assigned a value but never used.","Identifier","unusedVar","'bookCategory' is assigned a value but never used.","semi","Missing semicolon.","ReturnStatement","missingSemi",{"range":"24","text":"25"},[1025,1025],";"]
