+++ src/index.js
@@ -6,10 +6,10 @@ import store from './store';
 import App from './components/App';
 
 ReactDOM.render(
-  <React.StrictMode>
+  <React>
     <Provider store={store}>
       <App />
     </Provider>
-  </React.StrictMode>,
+  </React>,
   document.getElementById('root'),
 );
