+++ src/reducers/booksReducer.js
@@ -25,4 +25,5 @@ const booksReducer = (state = initialState, action) => {
     };
   }
 };
+
 export default booksReducer;
